+++ docs/CHANGELOG.rst
@@ -19,7 +19,8 @@ Added
   environment
 * Expose additional parameters in ``alignment-star``,
   ``cutadapt-single`` and ``cutadapt-paired`` processes
-* Add ``FastQC`` and ``SRA Toolkit`` to ``resolwebio/utils`` Docker image
+* Add ``FastQC``, ``igvtools``, and ``SRA Toolkit``
+  to ``resolwebio/utils`` Docker image
 
 Changed
 -------